resume-matcher/requirements.txt

# Core Framework
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5,<1.0.0

# NLP & ML
numpy>=1.23.0,<2.0.0
pandas>=1.5.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
spacy>=3.0.0,<4.0.0
openai>=1.0.0

# PDF Processing
pdfminer.six>=20221105

# Task Queue & Database
celery[redis]>=5.2.7
redis>=4.5.0
supabase>=2.0.0
